Move `observeRoomMemberIdentityStateChange` and associated classes to `libs:matrixui` module so they can be reused
