+++ libraries/matrixui/src/main/kotlin/io/element/android/libraries/matrix/ui/room/ObserveRoomMemberIdentityStateChange.kt
@@ -5,21 +5,22 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.features.messages.impl.messagecomposer
+package io.element.android.libraries.matrix.ui.room
 
 import androidx.compose.runtime.ProduceStateScope
-import io.element.android.features.messages.impl.crypto.identity.IdentityRoomMember
-import io.element.android.features.messages.impl.crypto.identity.RoomMemberIdentityStateChange
 import io.element.android.libraries.designsystem.components.avatar.AvatarData
 import io.element.android.libraries.designsystem.components.avatar.AvatarSize
 import io.element.android.libraries.matrix.api.core.UserId
+import io.element.android.libraries.matrix.api.encryption.identity.IdentityState
 import io.element.android.libraries.matrix.api.room.MatrixRoom
 import io.element.android.libraries.matrix.api.room.RoomMember
 import io.element.android.libraries.matrix.api.room.roomMembers
 import io.element.android.libraries.matrix.ui.model.getAvatarData
+import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filter
@@ -28,30 +29,33 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun ProduceStateScope<PersistentList<RoomMemberIdentityStateChange>>.observeRoomMemberIdentityStateChange(room: MatrixRoom) {
-    room.syncUpdateFlow
+fun MatrixRoom.roomMemberIdentityStateChange(): Flow<ImmutableList<RoomMemberIdentityStateChange>> {
+    return syncUpdateFlow
         .filter {
             // Room cannot become unencrypted, so we can just apply a filter here.
-            room.isEncrypted
+            isEncrypted
         }
         .distinctUntilChanged()
         .flatMapLatest {
-            combine(room.identityStateChangesFlow, room.membersStateFlow) { identityStateChanges, membersState ->
+            combine(identityStateChangesFlow, membersStateFlow) { identityStateChanges, membersState ->
                 identityStateChanges.map { identityStateChange ->
                     val member = membersState.roomMembers()
-                        ?.firstOrNull { roomMember -> roomMember.userId == identityStateChange.userId }
+                        ?.find { roomMember -> roomMember.userId == identityStateChange.userId }
                         ?.toIdentityRoomMember()
                         ?: createDefaultRoomMemberForIdentityChange(identityStateChange.userId)
                     RoomMemberIdentityStateChange(
                         identityRoomMember = member,
                         identityState = identityStateChange.identityState,
                     )
-                }
-            }
-                .distinctUntilChanged()
-                .onEach { roomMemberIdentityStateChanges ->
-                    value = roomMemberIdentityStateChanges.toPersistentList()
-                }
+                }.toPersistentList()
+            }.distinctUntilChanged()
+        }
+}
+
+fun ProduceStateScope<PersistentList<RoomMemberIdentityStateChange>>.observeRoomMemberIdentityStateChange(room: MatrixRoom) {
+    room.roomMemberIdentityStateChange()
+        .onEach { roomMemberIdentityStateChanges ->
+            value = roomMemberIdentityStateChanges.toPersistentList()
         }
         .launchIn(this)
 }
@@ -71,4 +75,15 @@ private fun createDefaultRoomMemberForIdentityChange(userId: UserId) = IdentityR
         url = null,
         size = AvatarSize.ComposerAlert,
     ),
+)
+
+data class RoomMemberIdentityStateChange(
+    val identityRoomMember: IdentityRoomMember,
+    val identityState: IdentityState,
+)
+
+data class IdentityRoomMember(
+    val userId: UserId,
+    val displayNameOrDefault: String,
+    val avatarData: AvatarData,
 )
